feat: improve dev workflow

+++ scripts/env.py
@@ -17,27 +17,33 @@
 # under the License.
 #
 import json
+import os
 import sys
 from pathlib import Path
 
-from sh import Command, CommandNotFound, RunningCommand
+from sh import RunningCommand
 
-from scripts.utils import Constants, detect_poetry
+from scripts.utils import Constants, detect_poetry, detect_pre_commit
 
 
 def configure() -> None:
     poetry = detect_poetry()
-    poetry("install", _out=sys.stdout, _err=sys.stderr)
+    poetry("install", "--sync", _out=sys.stdout, _err=sys.stderr)
 
-    # Generate pyright config
-    process: RunningCommand = poetry("env", "info", "-p", _err=sys.stderr, _return_cmd=True)  # type: ignore[no-untyped-def]
-    project_venv_path = Path(process.stdout.decode(Constants.ENCODING_UTF_8).strip())
-    project_venv_name = project_venv_path.name
-    venv_path = project_venv_path.parents[0].as_posix()
-    Constants.PYRIGHTCONFIG_PATH.write_text(json.dumps({"venv": project_venv_name, "venvPath": venv_path}))
+    # Generate pyright config for Zed
+    if os.getenv("ZED_TERM", "false") == "true":
+        process: RunningCommand = poetry(
+            "env", "info", "-p", _err=sys.stderr, _return_cmd=True
+        )  # type: ignore[no-untyped-def]
+        project_venv_path = Path(process.stdout.decode(Constants.ENCODING_UTF_8).strip())
+        project_venv_name = project_venv_path.name
+        venv_path = project_venv_path.parents[0].as_posix()
+        if not Constants.PYRIGHTCONFIG_PATH.exists():
+            Constants.PYRIGHTCONFIG_PATH.write_text(json.dumps({"venv": project_venv_name, "venvPath": venv_path}))
 
-    try:
-        pre_commit = Command("pre-commit")
+    # Skip pre-commit if not present
+    pre_commit = detect_pre_commit()
+    if pre_commit:
         pre_commit("install", _out=sys.stdout, _err=sys.stderr)
-    except (CommandNotFound, ImportError):
+    else:
         print("pre-commit isn't installed")
